requirements.txt
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
cairocffi==1.6.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cssselect2==0.7.0
Django==5.0.1
django-cleanup==8.0.0
et-xmlfile==1.1.0
fonttools==4.47.0
fpdf==1.7.2
freetype-py==2.3.0
html5lib==1.1
idna==3.6
lanshark-django-xhtml2pdf==0.0.7
lxml==5.0.0
numpy==1.26.3
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.4
pdfkit==1.0.0
pillow==10.2.0
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==3.17.4
pyphen==0.14.0
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.8
requests==2.31.0
rlPyCairo==0.3.0
segno==1.6.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.13
zopfli==0.2.3
